# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  profile.loginInput:
    properties:
      email:
        example: asrivatsa6@gmail.com
        type: string
      password:
        example: something123
        type: string
    required:
    - email
    - password
    type: object
  profile.profileOutput:
    properties:
      created_at:
        example: 1710579130
        type: integer
      email:
        example: asrivatsa6@gmail.com
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      name:
        example: abhinav
        type: string
      updated_at:
        example: 1710579130
        type: integer
      verified:
        example: false
        type: boolean
    type: object
...
  teammember.createTeamMemberInput:
    properties:
      admin:
        example: false
        type: boolean
      email:
        example: asrivatsa6@gmail.com
        type: string
      team_id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
    required:
    - email
    - team_id
    type: object
  teammember.profileOutput:
    properties:
      admin:
        example: false
        type: boolean
      created_at:
        example: 1710579130
        type: integer
      email:
        example: asrivatsa6@gmail.com
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      name:
        example: abhinav
        type: string
      updated_at:
        example: 1710579130
        type: integer
      verified:
        example: true
        type: boolean
    type: object
  teammember.teamMemberOutput:
    properties:
      admin:
        example: false
        type: boolean
      created_at:
        example: 1710579130
        type: integer
      profile_id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      team_id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      updated_at:
        example: 1710579130
        type: integer
    type: object
  teammember.updateTeamMemberInput:
    properties:
      admin:
        example: false
        type: boolean
      profile_id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      team_id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
    required:
    - profile_id
    - team_id
    type: object
...
  util.ErrorDetail:
    properties:
      location:
        type: string
      message:
        type: string
      value: {}
    type: object
  util.ErrorModel:
    properties:
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/util.ErrorDetail'
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: The comprehensive list of all Buggy APIs
  title: Buggy APIs
  version: "1.0"
paths:
...
  /profile/login:
    post:
      consumes:
      - application/json
      description: Login with email and password. Only verified users can login.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/profile.loginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Login with email and password.
      tags:
      - profile
...
  /profile/verify:
    get:
      description: Verify a profile. This link will only come from emails.
      parameters:
      - description: string
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Verify a profile.
      tags:
      - profile
  /team-member:
    post:
      consumes:
      - application/json
      description: Create a new team member. The profile and team has to exist already
        to add a member. The profile making this request must be an admin for the
        team already.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/teammember.createTeamMemberInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teammember.teamMemberOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Create a new team member.
      tags:
      - team-member
    put:
      consumes:
      - application/json
      description: Update a team members permissions.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/teammember.updateTeamMemberInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teammember.teamMemberOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Update a team members permissions.
      tags:
      - team-member
  /team-member/{team-id}:
    get:
      description: Get all team members.
      parameters:
      - description: Team ID
        in: path
        name: team-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teammember.profileOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Get all team members.
      tags:
      - team-member
swagger: "2.0"